Move command storage types under `command.storage` package
