+++ server/src/main/java/org/spine3/server/command/CommandStorage.java
@@ -34,9 +34,9 @@ import org.spine3.base.Commands;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
 import org.spine3.protobuf.TypeUrl;
+import org.spine3.server.command.storage.CommandStorageRecord;
 import org.spine3.server.entity.GetTargetIdFromCommand;
 import org.spine3.server.storage.AbstractStorage;
-import org.spine3.server.storage.CommandStorageRecord;
 
 import javax.annotation.Nullable;
 import java.util.Iterator;
