+++ server/src/main/java/org/spine3/server/storage/memory/InMemoryCommandStorage.java
@@ -27,7 +27,7 @@ import org.spine3.base.CommandStatus;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
 import org.spine3.server.command.CommandStorage;
-import org.spine3.server.storage.CommandStorageRecord;
+import org.spine3.server.command.storage.CommandStorageRecord;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
