+++ server/src/main/proto/spine/server/command/command_storage.proto
@@ -19,13 +19,13 @@
 //
 syntax = "proto3";
 
-package spine.server.storage;
+package spine.server.command;
 
 option (type_url_prefix) = "type.spine3.org";
 option (SPI_all) = true;
-option java_package = "org.spine3.server.storage";
+option java_package = "org.spine3.server.command.storage";
 option java_multiple_files = true;
-option java_outer_classname = "CommandStoreProto";
+option java_outer_classname = "CommandStorageProto";
 option java_generate_equals_and_hash = true;
 
 import "google/protobuf/timestamp.proto";
