+++ server/src/test/java/org/spine3/server/command/CommandStorageShould.java
@@ -35,8 +35,8 @@ import org.spine3.base.Commands;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
 import org.spine3.protobuf.AnyPacker;
+import org.spine3.server.command.storage.CommandStorageRecord;
 import org.spine3.server.storage.AbstractStorageShould;
-import org.spine3.server.storage.CommandStorageRecord;
 import org.spine3.test.Tests;
 import org.spine3.test.storage.ProjectId;
 import org.spine3.test.storage.command.CreateProject;
